python/04-okta-oauth-server/pyproject.toml
[project]
name = "mcp-incidents-okta"
version = "0.1.0"
description = "MCP server template — Streamable HTTP with Okta OAuth (Resource Server pattern)"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "pyjwt[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
